pyproject.toml
[project]
name = "ai-api-mcp"
version = "0.2.0"
description = "MCP server for integrating multiple AI APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.18.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.290",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ai-api-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]